Split CLI commands into separate files by resource (#98)
Move commands from monolithic Kaiten.swift into:
- SpaceCommands.swift (ListSpaces)
- BoardCommands.swift (ListBoards, GetBoard, GetBoardColumns, GetBoardLanes)
- CardCommands.swift (ListCards, GetCard, GetCardMembers)
- CustomPropertyCommands.swift (ListCustomProperties, GetCustomProperty)

Kaiten.swift retains only the entry point, GlobalOptions, and printJSON helper.

Closes #98
